Move `are_siblings` method to query trait
