+++ crates/fj-core/src/queries/mod.rs
@@ -11,8 +11,10 @@
 
 mod all_half_edges_with_surface;
 mod bounding_vertices_of_half_edge;
+mod sibling_of_half_edge;
 
 pub use self::{
     all_half_edges_with_surface::AllHalfEdgesWithSurface,
     bounding_vertices_of_half_edge::BoundingVerticesOfHalfEdge,
+    sibling_of_half_edge::SiblingOfHalfEdge,
 };
